modification MDP / indentation

+++ backend/controllers/comment.js
@@ -1,5 +1,6 @@
 const models = require("../models");
 const fs = require("fs");
+const jwt = require("jsonwebtoken");
 
 /* Création d'un commentaire */
 
@@ -14,19 +15,29 @@ exports.createComment = (req, res, next) => {
     commentObject.image = req.file.filename;
   }
   if (req.body.content == undefined && req.file == undefined) {
-    return res.status(400).json({ message: "Votre commentaire est vide !" });
+    return res.status(400).json({
+      message: "Votre commentaire est vide !"
+    });
   }
   console.log(commentObject);
   models.Comment.create(commentObject)
 
-    .then(() => res.status(201).json({ message: "Commentaire publié !" }))
-    .catch((error) => res.status(400).json({ error }));
+    .then(() => res.status(201).json({
+      message: "Commentaire publié !"
+    }))
+    .catch((error) => res.status(400).json({
+      error
+    }));
 };
 
 /* Modification d'un commentaire */
 
 exports.modifyComment = (req, res, next) => {
-  models.Comment.findOne({ where: { id: req.params.id } })
+  models.Comment.findOne({
+      where: {
+        id: req.params.id
+      }
+    })
     .then((comment) => {
       if (comment.User_id !== req.auth.userId) {
         res.status(403).json({
@@ -61,44 +72,64 @@ exports.modifyComment = (req, res, next) => {
           .catch((error) => res.status(400).json(error));
       }
     })
-    .catch((error) => res.status(500).json({ error }));
+    .catch((error) => res.status(500).json({
+      error
+    }));
 };
 
 /* Suppression d'un commentaire */
 
 exports.deleteComment = (req, res, next) => {
-  models.Comment.findOne({ where: { id: req.params.id } })
+  const token = req.headers.authorization.split(" ")[1];
+  const decodedToken = jwt.verify(token, "RANDOM_TOKEN_SECRET");
+  const isAdmin = decodedToken.isAdmin;
+  models.Comment.findOne({
+      where: {
+        id: req.params.id
+      }
+    })
     .then((comment) => {
-      if (comment.User_id !== req.auth.userId) {
-        res.status(403).json({ error: "Unauthorized request" });
-      } else if (comment.User_id == req.auth.userId) {
+      if (!isAdmin && comment.User_id !== req.auth.userId) {
+        console.log(isAdmin);
+        res.status(403).json({
+          error: "Unauthorized request"
+        });
+      } else {
         fs.unlink(`images/comments/${comment.image}`, () => {
           comment
             .destroy()
             .then(() =>
-              res.status(200).json({ message: "Commentaire supprimé !" })
+              res.status(200).json({
+                message: "Commentaire supprimé !"
+              })
             )
-            .catch((error) => res.status(400).json({ error }));
+            .catch((error) => res.status(400).json({
+              error
+            }));
         });
       }
     })
-    .catch((error) => res.status(500).json({ error }));
+    .catch((error) => res.status(500).json({
+      error
+    }));
 };
 
 /* Affichage de tous les commentaires */
 
 exports.getAllComments = (req, res, next) => {
   models.Comment.findAll({
-    where: { Post_id: req.params.id },
-    order: [["createdAt", "ASC"]],
-    attributes: ["id", "User_id", "Post_id", "content", "image", "createdAt", "updatedAt"],
-    include: [
-      {
+      where: {
+        Post_id: req.params.id
+      },
+      order: [
+        ["createdAt", "ASC"]
+      ],
+      attributes: ["id", "User_id", "Post_id", "content", "image", "createdAt", "updatedAt"],
+      include: [{
         model: models.User,
         attributes: ["id", "lastName", "firstName", "avatar"],
-      },
-    ],
-  })
+      }, ],
+    })
     .then((comments) => res.status(200).json(comments))
     .catch((error) => {
       res.status(500).json(error);
